+++ docker/web-entrypoint.sh
@@ -1,0 +1,6 @@
+#!/bin/sh
+set -eu
+
+python manage.py migrate --noinput
+python manage.py collectstatic --noinput
+exec waitress-serve --listen=0.0.0.0:8000 config.wsgi:application